chore(mobile): Refactors StreamList into Stream
