+++ apps/mobile/src/components/ModerationList/ModerationList.js
@@ -2,7 +2,7 @@ import React, { useCallback, useState } from 'react'
 import { View, FlatList, StyleSheet } from 'react-native'
 import { useNavigation } from '@react-navigation/native'
 import { gql, useMutation, useQuery } from 'urql'
-import { DECISIONS_OPTIONS } from 'components/StreamList/StreamList'
+import { DECISIONS_OPTIONS } from 'screens/Stream/Stream'
 import Loading from 'components/Loading'
 import ModerationListItem from 'components/ModerationListItem'
 import ListControl from 'components/ListControl'
