--- apps/mobile/src/components/StreamList/PostRow/PostRow.styles.js
@@ -1,11 +1,0 @@
-import { StyleSheet } from 'react-native'
-
-const streamMargin = 0
-
-export default StyleSheet.create({
-  postRow: {
-    paddingBottom: 12,
-    marginRight: streamMargin,
-    marginLeft: streamMargin
-  }
-})
