--- apps/mobile/src/components/StreamList/PostRow/index.js
@@ -1,3 +1,0 @@
-import component from './PostRow'
-
-export default component
