--- apps/mobile/src/components/StreamList/index.js
@@ -1,3 +1,0 @@
-import component from './StreamList'
-
-export default component
