+++ apps/mobile/src/screens/Stream/makeStreamQuery.js
@@ -158,3 +158,5 @@ const makeGroupPostsQuery = withChildPosts => gql`
   }
   ${postFieldsFragment}
 `
+
+export default makeStreamQuery
